fix: 모달 삭제 및 intercepting route 적용
