+++ src/app/(device-required)/@modal/default.tsx
@@ -1,0 +1,5 @@
+import { PropsWithChildren } from "react";
+
+export default function Default({ children }: PropsWithChildren) {
+  return children;
+}
