+++ src/app/(device-required)/@modal/layout.tsx
@@ -1,0 +1,11 @@
+"use client";
+
+import { PropsWithChildren } from "react";
+
+export default function Layout({ children }: PropsWithChildren) {
+  return (
+    <div className="fixed inset-0 z-50 flex items-center justify-center bg-black/50">
+      {children}
+    </div>
+  );
+}
