+++ src/app/(device-required)/@modal/pos/payments/sales/layout.tsx
@@ -1,0 +1,5 @@
+import RefLayout from "@/components/modal/RefLayout";
+
+export default function Layout({ children }: { children: React.ReactNode }) {
+  return <RefLayout>{children}</RefLayout>;
+}
