+++ src/app/(device-required)/pos/payments/sales/layout.tsx
@@ -1,0 +1,11 @@
+import { PropsWithChildren } from "react";
+import POSHeader2 from "../../_components/POSHeader2";
+
+export default function Layout({ children }: PropsWithChildren) {
+  return (
+    <div>
+      <POSHeader2 />
+      {children}
+    </div>
+  );
+}
